requirements.txt
requests
sqlalchemy
psycopg2-binary
pytest
pytest-sqlalchemy
icecream
pandas
